pyproject.toml
[tool.poetry]
name = "split-colors"
version = "0.1.0"
description = "Split colors"
authors = ["Felipe S. S. Schneider <schneider.felipe.5@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
split-colors = 'split_colors:run'

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.1.0"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
